If gltf material is BLEND, then always choose transparent material.
Previous version of the code would also check the alpha channel of pbr.baseColorFactor of the material and force opaque material if this value is 1.
In our case this behavior is "wrong", because colors (and thus alpha) are defined per-vertex, so pbr.baseColorFactor has its default value of (1,1,1,1).
So this would result in using the opaque material and render transparency with dithering, which is very low quality transparency.