AdvViz: basic support for glTF LINES primitive mode
ie list of segments: wouldn't be hard to support line loops and strips I
guess, but I'd need a dataset to test that.

No support for rendering with "attenuation" like point clouds: haven't
looked at the custom vertex factory and shader yet to see how hard it
would be to support lines.

Probably need to change the name of CesiumGltfPointsComponent and proxy,
unless it makes more sense to duplicate them.