Scene capture 2D: take the show-only / hidden list of actors into consideration
A scene capture component can be configured to render only a list of actors, or
to ignore a specific list of actors. Take those options into consideration to
avoid requesting a huge amount of useless data (depending on the position of
such capture components in the world) leading to filling the CPU and getting
awful performances.